tests/wpt/tests/encoding-detection/he-ISO-8859-8-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>he ISO-8859-8 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "ISO-8859-8", 'Expected ISO-8859-8');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/he-ISO-8859-8-late.sub.html");
</script>
